requirements.txt
# Core GUI framework
tkinter

# Image processing
Pillow>=10.0.0

# System tray
pystray>=0.19.0

# Syntax highlighting
pygments>=2.15.0

# HTTP requests and API clients
requests>=2.31.0
httpx>=0.24.0

# Async support
asyncio

# LLM API clients
openai>=1.3.0
anthropic>=0.7.0

# Input device handling (Linux)
evdev>=1.6.0

# Configuration and data handling
pyyaml>=6.0
toml>=0.10.0

# Development tools (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# System integration
psutil>=5.9.0